Time: 0 ms (100.00%), Space: 7.8 MB (38.30%) - LeetHub

+++ 0028-find-the-index-of-the-first-occurrence-in-a-string/0028-find-the-index-of-the-first-occurrence-in-a-string.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int strStr(string haystack, string needle) {
-        size_t pos = haystack.find(needle);
+        int pos = haystack.find(needle);
         if (pos != string::npos) { /*(find) if it found the string, will return first index that string start at else it will return (string::npos)*/
             return pos;
         } else {
